+++ languagetool-language-modules/pt/src/main/java/org/languagetool/rules/pt/PreReformPortugueseCompoundRule.java
@@ -36,10 +36,9 @@ public class PreReformPortugueseCompoundRule extends AbstractCompoundRule {
     super(messages,
             "Esta palavra é hifenizada.",
             "Esta palavra é escrita em conjunto.",
-            "Esta palavra é uma palavra ou com um hífen.");
-    super.setShort("Juntos grafias de palavras");
+            "Esta palavra é uma palavra ou com um hífen.",
+            "Juntos grafias de palavras");
   }
-
 
   @Override
   public String getId() {